booking/requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
drf-yasg==1.21.11
PyJWT==2.10.1
dj-database-url==3.0.1
requests==2.32.5
pika==1.3.2
django-cors-headers==4.9.0